+++ src/shared/src/types/network.rs
@@ -1,5 +1,55 @@
 //! Level 0 networks
 
+use std::collections::BTreeMap;
+
+use candid::{CandidType, Deserialize};
+
+use crate::types::Version;
+
+#[derive(CandidType, Deserialize, Clone, Debug, Eq, PartialEq, Default)]
+pub struct NetworkSettings {
+    pub enabled: bool,
+    pub is_testnet: bool,
+}
+
+#[derive(CandidType, Deserialize, Clone, Debug, Eq, PartialEq, Default, Ord, PartialOrd)]
+pub enum NetworkSettingsFor {
+    #[default]
+    InternetComputer,
+    BitcoinMainnet,
+    BitcoinTestnet,
+    BitcoinRegtest,
+    EthereumMainnet,
+    EthereumSepolia,
+    SolanaMainnet,
+    SolanaTestnet,
+    SolanaDevnet,
+    SolanaLocal,
+}
+
+#[derive(CandidType, Deserialize, Clone, Debug, Eq, PartialEq, Default)]
+pub struct TestnetsSettings {
+    pub show_testnets: bool,
+}
+
+#[derive(CandidType, Deserialize, Clone, Debug, Eq, PartialEq, Default)]
+pub struct NetworksSettings {
+    pub networks: BTreeMap<NetworkSettingsFor, NetworkSettings>,
+    pub testnets: TestnetsSettings,
+}
+
+#[derive(CandidType, Deserialize, Clone, Eq, PartialEq, Debug)]
+pub enum SaveTestnetsSettingsError {
+    UserNotFound,
+    VersionMismatch,
+}
+
+#[derive(CandidType, Deserialize, Clone, Eq, PartialEq, Debug)]
+pub struct SetShowTestnetsRequest {
+    pub show_testnets: bool,
+    pub current_user_version: Option<Version>,
+}
+
 pub mod marker_trait {
     use candid::{CandidType, Deserialize};
     use serde::Serialize;
